requirements.txt
geopy>=2.4.0            # Free geocoding (Nominatim/OpenStreetMap)
scrapfly-sdk>=0.8       # Rightmove scraping
jmespath>=1.0           # JSON parsing for Rightmove data
requests>=2.31
